Common DB-layer tests
